Fix permissions check for PUT/DELETE Resource Detail endpoints
[Closes #1428]

As commented in the issue we were not checking properly the permissions of
stakeholders and organizations assigned to a given resource as owner or
focal-points. The problem is related to the comparison of the results. We
have fixed that after a refactor to re-use these permissions check for
translations management, so now we have just get rid of the buggy and
duplicated code, to use the newer shared and fixed version of the code.