Add missing init. data in Get res. det. endpoint
[Closes #1482]

Added missing properties when returning resource data for initiatives, as we
have stored them in the DB, the same format we load them when getting all the
data from GET `/api/initiative/{id}` endpoint.

We have removed problematic `funding` property as it is not working well, to
load the raw `q35` property, that contains this information. In the future we
will refactor the code to provide a better meaning for the props of
Initiative resource type.