Fix GET suggested profiles endpoint
[Re #1488]

As explained in the issue now we can use any tag with the relationship of
`seeking` or `offering` (in the context of stakeholders). As such, it does
not longer make sense to keep using hardcoded relationship between offering
and seeking tags, as the same (any) tag can be used for both purposes, so we
have corrected the logic to match stakeholder profiles that have tags
of "offering" relationship that match the target stakeholder's tags
with "seeking" relationship and the other way around.

In order to avoid duplicated profiles, we have adapted the DB query to select
only distinct stakeholder data, since it is not needed to get tags related
info, since it is already reloaded from the profile db query before the data
is returned to FE. Then, as we are using UNION query to merge "seeking"
and "offering" kind of relationships-based profiles, we are sure we no have
duplicated profiles.

Fixed expectation of always receiving `page` param, so we now provide a
default value when it is not present, not to break the query.

Removed not used code, including an unrelated dependency to some `util` ns.

Now we also make sure we work with vectors in order to conj additional
profiles related to latest active users, so we first show the matching
profiles and then the rest, if needed. For the other sequences we are using
vectors since the response data need to realize the seqs anyway.