Take into account `limit` param in GET sugg. prof.
[Closes #1488]

Now we take into account FE-sent `limit` param in order to limit the number of
items per page for the get suggested profiles endpoint.

We have also refactored the way to expect and parse params to provide proper
typed validation and improve how we set the default values for those params.