Improve default values handling in get sug. prof.
[closes #1488]

Now we specify at Handler level (API layer) the default values for the GET
Suggested profiles endpoint, making the code simpler, as now we can ignore
transformations/coercions at deeper levels.

For this we have needed to remove the `optional=true` property of the target
params, since we need those params to be required in order to provide a
default value (which makes total sense). However, this is transparent for the
API user, since the params do not needed to be provided, as it is something
Malli and Reitit handle.