Fix URL handle in GET `/image/{image-type}/{id}`
[closes #1492]

We were not handling image URLs appropriately, so now we check first if we can
build a URL out of the image got from DB, so we try to download the image in
that case, as before. However, if we cannot make a URL we make sure we have a
valid string before trying to parse the image as Base64 content.

Now we also do proper logging, as we were not doing much before and we were
expecting the logger config in the wrong place as argument.

Now the fetch image function is also safer in case there is no `status` got
from the result, before trying to do further checks with its value.