Add excpt. hand. in GET `/image/{image-type}/{id}`
[closes #1492]

Now we wrap everything in a try-catch block in order to handle any exception
that might happen when getting an image, returning meaningful error to the
user and logging the exception as well.

We use `Throwable` class in order to catch any exception type, as `Exception`
does not cover all of them.