Fix img download by providing user-agent to req
[closes #1492]

Some avatar repositories need user-agent to be provided as part of the request
and we were missing that from BE. Now we inject it in the request as we fetch
it from the user's browser.

Documented the code a bit to explain the image handling flow depending on each
case, as it's a bit tricky.