[#1498] Refactor export tool and fix  entities' logos
* Moved all the export configuration to domain/export.clj.
* Added logic to create a link to the organisation's (or entity) logo
if it's a base64. This way it can be downloaded by accessing the URL
in the export file.
* Added an additional condition on the image handler to consider
organisation images stored in the database. This will change in the
future once we refactor the images management machinery. Ideally every
file/object should be stored in GCS and not the database.
* Closes #1498