Fix permission check when user didn't verify the email
* If the user doesn't verify the email our BE authentication code
assign a false value to the `:user` key in the autentication
details. That was causing a problem when checking for permissions as
we do `seq` on the `:user` key's value. That would case an
exception. So now instead of assigning a false value we assign an
empty map.
* I also changed in what order the permissions are checked because if
it's a read we don't have to check for the user as resources are
publicly readable.