Send email notif. in stakeh. reg. as last step
[closes #1502]

Right now we are sending email notifications when a new stakelholder is
registered, before the whole process has ended ok. So, for example, tags
creation related to the new stakeholder happens after the email notification
is sent.

However, we need to be sure all the steps related to the stakeholder
registration have been performed ok before sending the email notification,
which should be always the last step during the registration process.

So for doing the above, we have moved the notif. email sending place as last
step, rather than doing it when creating the stakeholder entity. Besides, we
do it once any possible exception has been thrown, so there is no need for
extra checks at this point.