Support new `source` property in all the resourc.
[Re #1508]

Added new `source` property for all the resource types, where we set the
default value to `gpml` enum, for existing resources or in case we explictly
do not set this property. This value means that GPML is the default source
platform of a given resource, but `cobsea` is the other supported value that
can be used too when setting a different `source` value for a given
resource.

We also support to provide this property to all the resource types across all
the endpoints for creating and editing resources. However, in order to
support this feature, we have needed to handle the transformations for the
Domain and DB layer in different ways, depending on the underlying code.
Ideally we would keep transformations of SQL-related parts in the DB layer,
but in several cases we have done it at the Handler (API) layer, in order to
reduce efforts. We will refactor the needed parts in the future.

For making sure the tests still work we have needed to manually add
`parameters` property to request, as we don't have the usual middlewares that
introduces coerced/default values for endpoint params there and the API
handlers expect those parameters to come from that side.

Updated tests for Project-related entity operations that needed the new
property to be there, as defaults injection does not work with tests, so we
need to provide always this property in this case.

Updated BRS API integration-related code to provide a default value for
`source` property in all the involved entities to be created/updated. We need
this in order to avoid issues, since we are considering all the declared
properties of each entity Domain, so we cannot just ignore `source`;
otherwise we would failing by trying to set this property with `nil` value,
as it never comes from the BRS API, but it is declared in the Domain layer.