Add case study related schema changes
[Re #1508]

Added needed tables as specified in the issue, following the current approach
as for the other resources.

In this case, we have kept just the relevant (actually used) enums values for
the case study associations, as in some cases we have old enum values, but
they cannot be removed, so the best is to add only the ones used currently in
other resources.