Updated Case Study-related schema
[Re #1508]

In order to simplify the re-usage of shared resources manipulation code, we
have renamed `created_at` property to be the same as in other resource types
(`created`) and we have also kept same pattern when naming columns for
Geo-Coverage-related tables.

Included `interested-in` among the rest of the resource relationship types, so
we can use Bookmark functionality as well as with other resources, despite
the fact that we will refactor that part in the near future.

Added forgotten new enum value for `topic_type`, that is needed to handle
permissions for this new resource type.