Support Creation of Case Study res. type
[Re #1508]

Added Create endpoint in the same way as we do with other resources, but
trying to make things better, by using a Schema defined in a new domain
layer, which is used to indicate which parameters are needed, which ones are
optional and the shape they have, as well as it serves to represent the
canonical entity.

In order to avoid further transformations of `source` and `geo_coverage_type`
enum properties when using this new POST endpoint, we have declared them as
keyword enums in the Schema, so we coerce them with Malli to keywords, so we
just need to later apply DB-related transformations. For doing this change we
have refactored too the other resources handling for the shared `source`
property. However, for `geo_coverage_type` we have transformed the enum value
options into keywords just for Case Study's schema, as otherwise we would
need to refactor lot of code, so we can do that later.

Fixed a bug when creating shared related content among the resources, as we
were not returning properly the affected rows after the creation of the
related content items. This was not affecting the other parts of the
application, since we were not checking the result of the query, but we do it
for Case Study, as an improvemenent in regards to other resources.

Updated allowed resource relationships in the domain.types related ns to be
up-to-date with the reality. Added some TODOs there as well for a future
refactor of the enum values.