Support Reviews, Submissions, Edit, List, Delete for Case Study
[Re #1508]

Changed the needed code to be able to assign reviewers to Case Study res.
types, see the submissions and be able to review and approve/deny a given
resource, as we do with the others.

We have also made the needed changes to take into account this resource when
loading and filtering content from GET endpoint of Browse API.

Besides, we also make sure we can edit and delete the resource from the
generic Details API endpoints.