Add new Contact endpoint for Cobsea
[Re #1517]

Added new endpoint similar to the Subscribe one, but in this case we are
sending contact form data.

We support only Cobsea contact form case (that is the only case for now), but
we have the endpoint ready for other cases, as long as we provide some
configuration about destination email address.

We have decided to make the `source` parameter mandatory, as it's a different
case as with other endpoints, since this would be mainly used from Cobsea and
better be explicit.