Add permissions metadata when creating events
[Re #1530]

Now we create a record of `rbac_context` table related to the new event
resource created, using the owner organisation's parent context as reference
if there is such a relationship. Otherwise, the parent context is related to
the application (root one). For this, we have exposed a generic method to be
re-used for the rest of the resources (but the location of the code might
change), using some constants to refer to well-known root and
organisation-related context type names. We do this as we only have a single
hierarchy level, to simplify things.

On the other hand, we use the `individual_connections` info to assign
different roles to the given users, under the respective contexts. For this
we filter the connections to only take care about the types related to
permissions, as the other ones are irrelevant here, but still we are
interested on saving that metadata (separated from permissions
infrastructure).

Left a TODO to decide if we want to keep storing redundant permissions
metadata, in the case the stakeholder has ownership above the organisation
owning the target resource.