Move new permissions-related code to its own ns
[Re #1530]

We have decided to put the new permissions system-related code into its
own namespace, under `service` folder. In that way, it's clear the
context when using those functions (for example, `resource` concept in
permissions context is a different thing).

Removed TODO comment related to the part of assigning roles to user when
creating an event, since we are ok with the approach, as it makes sense
the author of the event to be created is also the specific owner of
that resource, independently of the organisation that might be related
to that event.