Use new permissions system check for Create Event
[Re #1530]

Added a generic function to check permissions for the target operation
type, resource, context-type and user. This functions is flexible to
work with permissions check against specific context types or
application-level ones, including the possibility of asking for a
permissions that does not relate to a specific entity-type. We are only
using this function as part of Event Create operation, replacing
`approved-user`-related middleware by the usage of this function. In
this case, we are not using a middleware, since the current project
architecture is not ready to handle dynamic parameters to the
middlewares and we would ended-up with a tons of them to cover all the
different cases we have.

Refactored helper function to assign roles to a user in order to be
public, using a new name for the function calling it that uses
individual-connections info to build the user-roles assignments. In
this way, it is clearer now the purpose of each function.

Updated tests for creating event to also provide permissions meta-data
for the user that is being created as part of the fixture, in order to
have the right permission to be able to create the event resource.

Refactored `gpml.handler.event` namespace to use
`gpml.handler.responses` instead of the Ring-related version, as we are
doing it with other parts of the code.