Fix migration of user permissions for admins
[Re #1530]

We are not interested in giving full-permissions access to admin-role
users that are not approved yet or they have been rejected, so have
changed that accordingly.

On the other hand, admins should also have `approved-user` role when
they are migrated from previous permissions system, just in case they
keep those permissions if they become later to regular users.