Refactor approach convert orgs to member
[Re #1530]

Refactor approach convert orgs to member

[Re #1530]

Instead of deleting non-member organisation and re-creating it as member
now we just expose a different endpoint for doing the operation as an
update over the existing organisation. In this way, the code is much
simpler as we can keep current relationships of the target organisation
and we just need to update a flag (property) of it. Besides, it was a
mess to have this operation as part of stakeholder "creation" handler.

We have needed to change the DB query and handler to let `is_member`
property to be changable, so as such we have left also a TODO in the
update handler, so we skip allowing setting this flag, as we don't this
flag to be changable at any moment.