Fix update stakeholder with new org and add permissions metadata
[Re #1530]

We were not disocc-ing the org id when creating it, as FE is sending -1
value to indicate it is a new organisation, so now we do it to avoid
using that id when creating the new organisation.

After creating the new organisation we also assign the right ownership
roles to the creator user.

There is a little comment about handling ownership changes, since we
cannot control the user removes the organisation from his profile.