Check Rbac permissions when creating a case-study
[Re #1530]

Now we are checking if the user has the right permission to create a
case-study, as we do with Event creation already.

As we are using the new permissions system to check the permission we
now have removed the approved-user-related middleware, since it is not
needed anymore in this context.