Fix project's and comment's rbac permissions
[Re #1530]

We needed to add `project/read` new permissions, assigned to
`approved-user` role, as projects cannot be read publicly as other
resources.

On the other hand, Comment entity related permissions have been removed
since only their owners (and maybe admins) will be able to update or
delete them, so no need to have a context type for this functionality
and neither have specific permissions other than the one needed to
create comments, that are related to the `application` context type.