Fix context hierarchy assignment and add delete resource fn
[Re #1530]

We were not adding some contexts to the right parent context. We were
using `nil` reference when `application` one should be used, as all the
resources that do not belong to `organisation` context type should be
hanging from `application` one, which is the actual root context type.

On the other hand, we were missing a utility function to delete resource
contexts, so now we can use it.