Instrument Project API
[Re #1530]

We have added the needed permissions check for each operation of the
Project API endpoint, removing the not needed middlewares.

On the other hand, we have removed the `stakeholder-id` property when
receiving the data to be updated from a Project entity, since that does
not make sense, as the owner of the Project is the only one able to
edit a project and it does not make sense to transfer the ownership of
it.

Renamed db transaction binding to better identify the variable and avoid
mistakes.