Admins: ensure `resource-owner` role as fallback
[Re #1530]

Now we make sure the author of the resource being created has
`resource-owner` role over the resource, whenever no explicit
individual connections are provided. This is applied to all the
resource types.

Besides, we have simplified some parts of the code that did not needed
an extra DB query to fetch user data related to a given resource, as
that data is already injected by a middleware.