Add missing permission to list assigned reviews
[Re #1530]

Right now only approved users or super-admins can list their assigned
reviews. So we have needed to create a new permission for it, assigning
it to the `approved-user` role only. That means that the user needs to
be approved even if he has assigned reviews in order to list them, as
that is the original behaviour that needs to be kept.

We still need to instrument the code to use the permission.