Instrument Review API with RBAC
[Re #1530]

Now all the related endpoints have needed permission checks, removing
previous way to check permissions.

On the other hand, each operation that needs to assign or unassign roles
has been instrumented as expected.

We have needed to make public a function to tranform entity types to
context types, transforming underscores to dashes in the process. As we
were calling that function with keywords as well, we now make sure it
gets always a string as input.

Regarding previous permissions checks for reviewing a particular
resource, now we have replaced some 403 responses by 404 ones, as in
those contexts the permissions have been already checked and what it
could happen is that we cannot find the requested review information,
at most. Maybe there is no even need to check that, but it was simpler
to keep the code like that.