Fix context type param in Detail API with RBAC
[Re #1530]

We were not taking into account that the defined RBAC context types use
dash-case instead of underscores, so we have needed to apply the same
transformations we have been doing for the instrumented Review API. In
this case, the generic permission check function can accept a string
that now is trasnformed before using it as rbac context for the
permission check.

Left a TODO to remember deleting resource contexts when deleting the
resource itself, just in case.