Fix LEAP API's `plasticToolkitTags` handling
[Re #1536 #1530]

As explained in the issue, we have adapted the code in regards to the
breaking change, by splitting the joined tags by comma characters,
trimming each extracted content as well. As a side note, the code that
deal with this part is generic, since it handles different cases to
extract tags, so we check if we are dealing with a string to be sure we
can split the tags beforehand.