Instrument LEAP API import with RBAC
[Re #1530]

Now we generate the needed RBAC Contexts when creating new tag and
policies during the import process. The way is done keeps the process
optimized to perform multiple inserts in the DB and minimize the
refactor needed, while it keeps the consistent behaviour of aborting
the whole process if something fails, except for the tags, that are
created anyway (in case their RBAC contexts can be created as well),
which is consistent with the previous behaviour as well.