Instrument BRS API import with RBAC
[Re #1530]

Now we generate the needed RBAC Contexts when creating new tag and
different resources during the import process. The way is done keeps
the process optimized to perform multiple inserts in the DB and
minimize the refactor needed, while it keeps the consistent behaviour
of aborting the whole process if something fails, as the operation
relies on a transaction that is shared among the rest of the
operations.

Left a TODO in case we want to extend the supported entity types to
import from BRS, since we are not converting the `entity-name` to the
RBAC `context-type-name` fully, as we are not replacing possible
underscores, but at this point it is not really needed.

As a side note, there is a lot of noise just because we need to pass the
logger deep down all over the places. In the future we should improve
this.