Add featured order by to browse API
[Re #1537]

* The 'featured' order by parameter will order in DESC mode as
PostgreSQL booleans are 0 and 1 at the of the day. This way we always
get the featured resources first in the list. In order to avoid the
list being chaotic, subsequent resources will be ordered by
'created'. So, we at least keep some meaningful order of resources
that are not featured.