Finish instrum. `/api/stakeholder/{id}` PATCH API
[Re #1530]

Now we reflect the role changes in RBAC by adding the user to the RBAC
super-admins list whenever the "ADMIN" role is assigned to him or
removing that user from being a super admin whenever he loses
the "ADMIN" role.

We have kept the current roles assignment as it is not to break the API
for FE. In this case, the error response might be different now since
we have a single transaction to ensure the process is completed ok as a
whole, so the error message might be different depending on the failing
point. But this looks not important for FE, since it doesn't show any
specific error.

In order to add or remove super-admins in RBAC we have exposed two new
methods in the permissions-related service.