Instrument Tag API with RBAC
[Re #1530]

Instrumented code to use RBAC to check permissions when needed. In this
case no other change was needed.

As a side note, the modified PUT API endpoint for updating tags do not
work properly, but it should be fixed in a separated task, as it is
totally unrelated to the permissions refactoring task.