Fix permissions check on Detail GET API endpoint
[Re #1530]

Now we are handling the permissions over non-published resources
appropriately. Thus, instead of letting the endpoint to be consumed
whenever the resource type is not `organisation` or `stakeholder` type,
now we check if its status is not `APPROVED` to do further permissions
check (`x/read-draft` kind), so only users with `resource-owner` or
`resource-editor` role can have access to the endpoint, in that case.

Fixed the comment about checking the permissions for `stakeholder` or
`organisation` resource type, since the organisation-related part was
being omitted.