Finaly fix Details GET API permissions
[Re #1530]

The previous code was wrong not providing the resource id as parameter
and now we do like with Organisation's GET API endpoint, we check for
the permission to read organisation data at root context level, where
is expected to be for any approved or non-approved user.

Improve formatting of a file as well.