Fix organisation focal-points permissions (un)assignment
[Re #1530]

Organisation association are very tricky because they have
the `focal-point` system. If organisation A has focal point
Z, and organisation B has focal-point Z and Y it can lead
to duplicates and removal of permissions from focal points
when there shouldn't be. There are two problems here in the
context of a specific resource F:

1 - When stakeholder Z is assigned as focal point of both
org A and B, at time of assigning them the RBAC permissions
we'll encounter two role assignments with the same role,
user and context (for resource F). Because both orgs have
the same focal-point. To solve that we remove the
duplicates and assign the permission once.

2 - If we remove org A from the resource association,
stakeholder Z should lose their permission on resource F
BUT org B is still associated as owner of resource F and
stakeholder Z is a focal-point of org B. Therefore, we
cannot remove its permissions.

These are very extreme use cases where one stakeholder is a
focal-point of two different organisation that are
associated as owners of a resource.