Add thread transactions mechanism for executing seq of actions
[Re #1543]

* With this we can specify a vector containing N fns to execute in
sequence. Each fn needs to signal either success or failure using a
map like so:

`{:success? true|false}`

Extra details in the map are allowed and recommended as they are passed
as the context for the next function in the sequence. Each fn node can
have a rollback function associated with it in case the next fn in the
sequence fails. Rollback functions are not mandatory and you can use
them when needed. In the following commits there will be a full
example of this mechanism in the new file service.