Add mapping from mime-types to common-extensions
[Re #1543]

* We weren't storing the right extension for the file in some
cases. The mime-type's suffix is not always the same as the extension
name. for the most common mime-types such as `application/pdf`,
`image/jpeg` and `image/png`, the extension of the file can be easily
inferred by take the suffix of the mime-type (the string after the
slash), but for other like docx files it's completely different. So,
we are mapping those that are used in GPML to make sure we store the
right extension name.