Impl files-migrator API to programmatically migrate files
[Re #1543]

* This API endpoint will serve to trigger file migration for certain
entities. These entities are `event`, `case_study`, `stakeholder` and
`organisation. These entities' files can't be migrated using a
database migration because some of records store files in a
different. Some store files as base64 or public URLs (not pointing to
GCS). So we have to go row by row and handle each case and finally
upload the file to GCS.
* This API will lose its purpose once the migration is done as well as
the associated hugsql function. In which case we can safely delete
everything.