Fix file management refactor DB migration
[Re #1543]

We needed to take into account dash characters among the
allowed/expected file names when building the temporal image tables for
the "image" and "thumbnail" properties during the DB migration related
to already stored GCS images for the different resource types.
Otherwise, the file name string generation ends up with an empty value,
which violates the NOT NULL constraint in the File table's Name
column.

After further checking the code and both Prod and Test Env DBs, we are
confident that these are the only changes needed to fix the issue and
avoid any other surprises (since we know how the code is building the
image urls when uploading them to GCS).