Fix access to get private file URLs (#1565)
[Re #1543]

We were not including the namespace that extends the GCSStorageClient
protocol with the methods to get blob signed URL and delete blob-relate
functionality, that was preventing the reading of private image files
from GCS. This was not reproduced locally, since there we load always
all the namespaces, which is not the case with the Uberjar. So now we
require the mentioned file as part of the File service, that is
required already, fixing the issue.