Fix image update in resource edition
[Re #1543]

We were using the inverse logic to handle the result of deleting the
image of a resource that was going to be replaced by another image as
part of the update process. The main problem is that we were throwing
an exception to rollback the DB-related updates on the resource data
when the old image delete process was successful. However, we want the
opposite and throw the exception only when the process goes wrong and
continue with the DB-related updates otherwise. Now we have fixed that
logic in all the affected places.

Tests have been updated as well to avoid passing explicit nil `image`
property when updating resources, since in that case we are expecting a
previous old image when we do not have it for those cases.