Fix limbo image state when failing res. update
[Re #1543]

In the case of failing the process of updating a resource, when updating
images, we could end-up keeping deleted image referemces from Object
Storage (GCS), in the DB, as a result of rolling back the DB-related
operation. Hence, if we end-up in that case, we try to delete those
files from the DB, so at least it's clearer for the user and no broken
links are displayed. This is the simples approach, since rollbacking
fully the changes to re-create the files would imply more complexity
and resources consumption for the server and it is not worth it, since
the user wanted to replace those files anyway, so he can just try it
again by uploading the new files in newer updates.

This process has to be improved since we are doing the mentioned
operations as part of a Catch exception block, so this code is not
catched in another block.