Fix stakeholder creation service
[Re #1576]

We were not getting the result right from the roles assignment utility
function, when we needed to create `resource-owner` role for created
non-member organisation related to the stakeholder being created. Now
we have fixed that.

Besides we also have removed a print of a exception catched, in a
separated ns related to organisations.