Fix and improve `/api/profile/suggested` GET API
[Re #1578]

Improved the response of this API endpoint to always return an empty
vector as value of `suggested-profiles` key, instead of returning a
collection of empty maps, which was kind of cumbersome to deal with for
the frontend. For this we have made the `get-stakeholder-profile`
related function to return `nil` instead of an empty map, when no
profile has been found, so we can filter out those results from the
collection of suggested profiles. In this case, we have kept the
original response of an empty map in the expected
(single-profile) use-cases.

On the other hand, we were wrongly using `get-stakeholder-profile`
function, passing wrong arguments to this function within this
endpoint's handler. So now we have fixed that and we get the expected
results as well.