Impl. `/api/organisation{id}/request-membership` API
[Re #1580]

Refactored `to-member` endpoint as explained in the issue to make the
user to request memership of organisation by updating its
`review_status` to `SUBMITTED` instead of being a direct process, so a
GPML admin can review the process.

We have also made sure we update `affiliation` property if there is an
organisation linked to the target stakeholder, as that was only set in
some special cases that no don't make sense anymore.

We also make sure that the target entity is not under `REJECTED` status
or it is a member already to avoid strange situations.

In order to update the `review_status` we have modified a common DB
query, but disoccing that field from the places where we do not want
this to be set from the users.