Fix experts invitations related permissions handling
[Re #1582]

Fixed query to get experts results needed to be able to detect an expert
user needs to be converted to an approved user.

Besides, we have instrumented the code of update stakeholder-related
function to unassign unapproved user role from the expert user in case
it is detected as such, and be able to assign the approved user role in
that case as well. The added functions have their rollbacks as well and
we have also added one for the update stakeholder's final function, to
have a consistent behaviour.

On the other hand the endpoint to accept invitations looks not to be in
use and such this looks like the only entrypoint to fix, but we will
check.

Now we are also catching better the exceptions, using Throwable class.