Avoid duplicate tags errors, using existing tags
[Re #1588]

As explained in the issue, we want to avoid throwing duplicate tag
errors and use the tags that exist when needed, making it transparent
for the frontend. So in order to achieve that we avoid conflicts when
we find duplicate tags during the creation process, returning the data
of both newly created tags and the conflicting ones, making it
transparent for the frontend.

Once we have tried to create all the tags, we try to fetch related rbac
context to know if we need to create a new context or not, since we
cannot just try to create it, as it would cause an SQL error that could
abort any ongoing DB transaction.

Ideally the frontend should send the ids when we are dealing with
existing tags and we should change the backend code to avoid trying to
create already existing tags, but with this approach we tackle all the
issues at the same time and we avoid throwing errors in those cases,
making the code simpler as well.

Removed also an outdated unrelated comment found.