Update migration 193 to add chat_account_username column
[Re #1590]

* The RocketChat account creation process is delayed until a GPML
admin approves an user. But to create the account we require a
username, which should be an user input. So the idea is that during
registration, if the user opt-in to use the chat, they should also
input which username they want to use. The FE will call the BE to
check whether that username is free to use. This way we it's easier
for to determine if an username is free or not and also reserve that
username for later when the user needs to be created.