Remove non-approved users from prof. suggestions
[Re #1597]

As explained in the issue we have needed to remove the non-approved
stakeholders from the results of profile suggestions. In this case, we
have applied the filter in two different DB queries, including one to
get latest activity from stakeholders, where it was not sensible to
make this filter optional.