Refactor `geo-coverage` for Initiatives
[Re #1615]

Now we are dealing with `geo-coverage` in the same way as with the other
resource types. From one side, `geo_coverage_type` property has been
added to the Initiatives DB model and the related data that was stored
in other format and property has been migrated there. There are some
non-used values for the `geo_coverage_type`, but we can just replace
those the next time those initiatives are updated, to keep things
simple.

On the other hand, we have gotten rid of the old `country` and
`subnational_city` legacy properties that have been replaced already by
the shared `geo-coverage` concept, dropping the related columns about
it, that also included `geo-coverage` values-related duplicated
information. Besides, legacy DB views related to initiatives have been
deleted, as they were relying on those old properties and we could not
drop them otherwise.

Some tests have been also removed as they were as legacy as the
refactored code (which implies the rename and modifications of some
example data files used there) and some others have been updated to
adapt them to the refactored code or just simplify the logic that was
unnecesary complex.

When adapting BRS integration-related code, we have also unified the way
to transform data before being persisted, to avoid doing operations
twice and differently in separated places.

As a side note, there is an endpoint for getting initiative data that
does not use the common `geo-coverage` related tables for providing
this information. As it was using legacy `country` and similar
references, we have get rid of this logic. However, its future is being
discussed, as it should not be needed this endpoint at all.